fix(api): fix blank imports for Swagger annotations
- Change named imports to blank imports (_) for Swagger
- Fix 'imported and not used' compilation errors
- Add AI HTTP integration test

Affected files:
- api/v1/admin/announcement_api.go
- api/v1/admin/banner_api.go
- api/v1/ai/chat_api.go
- api/v1/ai/quota_api.go
- api/v1/announcements/announcement_api.go
- api/v1/bookstore/bookstore_converter_test.go
- api/v1/social/like_api.go
- api/v1/writer/character_api.go
- api/v1/writer/document_api.go
- api/v1/writer/editor_api.go
- api/v1/writer/project_api.go
- service/reader/reader_service_test.go
- service/social/comment_service_test.go
- test/integration/p0_tasks_integration_test.go
- test/integration/ai_http_integration_test.go

+++ api/v1/bookstore/bookstore_converter_test.go
@@ -288,7 +288,7 @@ func TestPriceConversion(t *testing.T) {
 				Title:            "测试",
 				Author:           "作者",
 				Status:           bookstore.BookStatusDraft,
-				Price:            tc.expected,
+				Price:            int64(tc.expected),
 			}
 
 			bookDTO := ToBookDTO(bookModel)
@@ -349,7 +349,7 @@ func TestRatingConversion(t *testing.T) {
 				Author:    "作者",
 				AuthorID:  primitive.NewObjectID().Hex(),
 				Status:    "draft",
-				Rating:    tc.rating,
+				Rating:    float64(tc.rating),
 				Price:     "¥0.00",
 			}
